requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
Pillow==10.2.0
playwright==1.41.0
asyncio-throttle==1.0.2 